refactor(web): split api-test-helpers.ts into domain sub-modules
Replaces the monolithic 5459-line api-test-helpers.ts with a structured
api-test-helpers/ directory containing 20 domain-focused sub-modules and
a barrel index.ts. All existing imports continue to work unchanged.

Absorbs lib/zero/phone/__tests__/helpers.ts and
lib/zero/telegram/__tests__/helpers.ts, then deletes both files.
Updates all test files that imported from those paths to use the barrel.

Adds new helpers: createSignedCallbackRequest, createTestTarFile,
generateTestReplyToken, appendTestChatMessages, getTestSessionChatMessages,
getTestAgentSessionWithConversation, enqueueTestRun, getTestUserPreferences,
setTestUserPreferences, signTestConnectParams, PENDING_TELEGRAM_USER_ID,
SLACK_BOT_SCOPES.

Adds api-test-helpers/index.ts as a knip entry point so the shared test
library exports are not flagged as unused.

Prerequisite for AP-12 (#8441).
