+++ turbo/apps/web/src/__tests__/api-test-helpers/index.ts
@@ -1,0 +1,22 @@
+// Barrel re-export for api-test-helpers sub-modules.
+// All existing imports from api-test-helpers continue to work unchanged.
+export * from "./core";
+export * from "./auth";
+export * from "./org";
+export * from "./agents";
+export * from "./runs";
+export * from "./schedules";
+export * from "./storage";
+export * from "./secrets";
+export * from "./connectors";
+export * from "./callbacks";
+export * from "./email";
+export * from "./slack";
+export * from "./github";
+export * from "./telegram";
+export * from "./phone";
+export * from "./credits";
+export * from "./users";
+export * from "./skills";
+export * from "./runner";
+export * from "./export";
