+++ turbo/knip.json
@@ -12,7 +12,8 @@
         "navigation.ts",
         "scripts/**/*.ts",
         "**/__tests__/**/*.{test,spec}.{ts,tsx}",
-        "**/*.{test,spec}.{ts,tsx}"
+        "**/*.{test,spec}.{ts,tsx}",
+        "src/__tests__/api-test-helpers/index.ts"
       ],
       "project": ["**/*.{ts,tsx}", "scripts/**/*.ts"],
       "ignoreDependencies": [
